refactor(cli): migrate model-provider command tests to colocated structure
Complete the migration of model-provider tests to colocated structure:
- Add set-default.test.ts with success, validation, and error tests
- Add index.test.ts for parent command help text test
- Delete src/__tests__/model-provider-command.test.ts

The setup.test.ts and delete.test.ts were already migrated in previous PRs.
